src/requirements/req_pindock.txt
BeautifulSoup==3.2.0
Django==1.3.1
Genshi==0.6
Markdown==2.0.3
MySQL-python==1.2.2
PAM==0.4.2
PIL==1.1.7
-e git://github.com/pinax/pinax.git@c90f20ef9e5ed3a890cb2524634075477b7d35fc#egg=Pinax-dev
South==0.7.3
Trac==0.12.2
Twisted==10.1.0
Twisted-Conch==10.1.0
Twisted-Core==10.1.0
Twisted-Lore==10.1.0
Twisted-Mail==10.1.0
Twisted-Names==10.1.0
Twisted-News==10.1.0
Twisted-Runner==10.1.0
Twisted-Web==10.1.0
Twisted-Words==10.1.0
akismet==0.2.0
apt-xapian-index==0.41
chardet==2.0.1
-e git://github.com/chrisglass/cmsplugin_facebook.git@1316e91a40628752930cbc70de2804207d7e314b#egg=cmsplugin_facebook-dev
cmsplugin-filer==0.8.0
cmsplugin-markup==0.1-alpha-4
cmsplugin-news==0.3c
cmsplugin-portfolio==0.1
configobj==4.7.2
distribute==0.6.10
django-announcements==0.1.2
django-appmedia==1.0.1
django-autofixture==0.2.5
django-blog-zinnia==0.8
-e git://github.com/ojii/django-classy-tags.git@7353938c61269962c717ed5043c229b93508172e#egg=django_classy_tags-dev
django-cms==2.1.3
django-compressor==0.7.1
django-countries==1.0.5
django-dajaxice==0.2
django-db-log==2.2.1
django-debug-toolbar==0.8.4
django-devserver==0.1.7
django-email-confirmation==0.2.dev6
django-filer==0.8.2
django-mailer==0.2a1.dev3
django-mptt==0.3.1
django-mptt-2==0.3.0
django-notification==0.2a1.dev4
django-openid==0.2.3.dev1
django-pagination==1.0.7
django-paging==0.2.4
-e git://github.com/ashchristopher/django-sass.git@0c77ae0702e2fc771a2ec83a1cb36f99ba726451#egg=django_sass-dev
django-sekizai==0.4.2
django-staticfiles==1.0.1
django-tagging==0.3.1
django-templatetag-sugar==0.1
django-timezones==0.2.dev2
django-tinymce==1.5.1a2
django-uni-form==0.7.0
django-xmlrpc==0.1.3
easy-thumbnails==1.0-alpha-18
httplib2==0.7.1
idios==0.2a1.dev2
lazy-paginator==0.1
markdown2==1.0.1.17
mercurial==1.6.4
mod-python==3.3.1
oauth2==1.5.170
pyOpenSSL==0.10
pyasn1==0.0.11a
pycrypto==2.1.0
pyinotify==0.8.9
pyparsing==1.5.5
pyserial==2.3
python-apt==0.7.100.1-squeeze1
python-debian==0.1.18
python-openid==2.2.5
pytz==2011h
reportbug==4.12.6
sorl-thumbnail==11.01
textile==2.1.4
virtualenv==1.6.4
virtualenvwrapper==2.8
wsgiref==0.1.2
yolk==0.4.1
zope.interface==3.5.3
